test: 회원에서 사용하는 Store,Product 예약내역 DTO의 디렉토리 위치 변경 및 필요없는 DTO 삭제
